⚡refactor and steamline codebase
