+++ bot/time-handling.js
@@ -1,0 +1,7 @@
+// this is where the logical decision making based on the time triggered will sit
+
+const isOnTheHour = () => new Date().getMinutes === 0;
+
+exports = {
+  isOnTheHour,
+}
